requirements.txt
Django==1.7.1
Pillow==2.6.1
django-custom-user==0.5
djangorestframework==2.4.3
wsgiref==0.1.2
